Refactor anti-hallucination module into modular components with backward compatibility
