+++ REFACTORING_PLAN.md
@@ -62,17 +62,17 @@ This document outlines the incremental refactoring plan for the LLM-RAG codebase
   - [x] Implemented extensive test coverage
   - [x] Maintained backward compatibility with WebPageLoader alias
 
-## Phase 5: Anti-Hallucination Refactoring
+## Phase 5: Anti-Hallucination Refactoring (Completed)
 
 - Break down `anti_hallucination.py` (694 lines) into focused modules:
-  - [ ] Create `src/llm_rag/rag/anti_hallucination/` directory
-  - [ ] Extract entity verification to a separate module
-  - [ ] Extract similarity-checking to a separate module
-  - [ ] Extract post-processing to a separate module
+  - [x] Create `src/llm_rag/rag/anti_hallucination/` directory
+  - [x] Extract entity verification to a separate module
+  - [x] Extract similarity-checking to a separate module
+  - [x] Extract post-processing to a separate module
 - Improve configurability and extensibility:
-  - [ ] Make verification strategies pluggable
-  - [ ] Enable runtime selection of strategies
-  - [ ] Add configuration validation
+  - [x] Make verification strategies pluggable
+  - [x] Enable runtime selection of strategies
+  - [x] Add configuration validation
 
 ## Phase 6: Testing and Documentation
