+++ src/llm_rag/rag/resources/stopwords_en.json
@@ -1,16 +1,1 @@
-[
-  "a", "an", "the", "and", "or", "but", "if", "then", "else", "when",
-  "at", "from", "by", "for", "with", "about", "against", "between",
-  "into", "through", "during", "before", "after", "above", "below",
-  "to", "up", "down", "in", "out", "on", "off", "over", "under",
-  "again", "further", "once", "here", "there", "where", "why", "how",
-  "all", "any", "both", "each", "few", "more", "most", "other",
-  "some", "such", "no", "nor", "not", "only", "own", "same", "so",
-  "than", "too", "very", "s", "t", "can", "will", "just", "don",
-  "should", "now", "this", "that", "these", "those", "i", "you", "he",
-  "she", "it", "we", "they", "me", "him", "her", "us", "them", "my",
-  "your", "his", "its", "our", "their", "mine", "yours", "hers", "ours",
-  "theirs", "am", "is", "are", "was", "were", "be", "been", "being",
-  "have", "has", "had", "having", "do", "does", "did", "doing", "would",
-  "could", "should", "might", "may", "must", "shall", "will", "can"
-]
+["a", "an", "the", "and", "or", "but", "if", "then", "else", "when", "at", "from", "by", "for", "with", "about", "against", "between", "into", "through", "during", "before", "after", "above", "below", "to", "up", "down", "in", "out", "on", "off", "over", "under", "again", "further", "once", "here", "there", "where", "why", "how", "all", "any", "both", "each", "few", "more", "most", "other", "some", "such", "no", "nor", "not", "only", "own", "same", "so", "than", "too", "very", "s", "t", "can", "will", "just", "don", "should", "now"]
